+++ CHANGELOG.md
@@ -111,6 +111,7 @@
 * sdk: add `Client::try_connect_relay` ([Yuki Kishimoto])
 * sdk: add `Client::wait_for_connection` ([Yuki Kishimoto])
 * relay-builder: custom http server ([v0l])
+* cli: allow setting a port in `serve` command ([Yuki Kishimoto])
 
 ### Fixed